requirements.txt
requests>=2.28.0
beautifulsoup4>=4.12.0
pandas>=2.0.0
plotly>=5.0.0
matplotlib>=3.6.0
pyyaml>=6.0
python-dateutil>=2.8.0
pytest>=7.0.0
pytest-cov>=4.0.0
